Require Authlib 0.10.
The bizarre problem occurred as a result of a version mismatch, but going to
the latest version solves it. So let's mandate that. NOTE: I am NOT demanding
a specific version here. The problem is unlikely to recur, and 0.11 probably
will be fine (when it comes out). Mandate as little as is necessary.