Apply category changes via both Kraken and Helix.
There've been some weird issues with VODs not seeing a category change, and
some channel bots (StreamElements, Nightbot?) also not seeing it. Test this
and see if the problems are gone. And then some time in the future, it'd be
really nice if we didn't need this pile of hacks... https://xkcd.com/1479/