Fix a bug that's probably not bothered anyone, ever...
In the rare situation where someone sends a text frame that isn't valid JSON,
the intended result was to ignore the frame; but it would probably have sent
back a 500 or something and closed the connection. It's a protocol error in
any case, so I really don't care, but it's nice to have it right.